+++ pipelines/capture__gtfs/CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Adicionado
 
 - Adiciona o teste `test_check_trajeto_alternativo__trips_gtfs` ao flow `capture__gtfs`
-- Adiciona o teste `test_formato_evento__ordem_servico_trajeto_alternativo_sentido` ao flow `capture__gtfs`(https://github.com/RJ-SMTR/pipelines_v3/pull/411)
+- Adiciona o teste `test_formato_evento` em `ordem_servico_trajeto_alternativo_sentido` (planejamento e staging) ao flow `capture__gtfs` (https://github.com/RJ-SMTR/pipelines_v3/pull/411)
 
 ## [1.2.0] - 2026-06-23