+++ queries/models/planejamento/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Adicionado
 
-- Adiciona o teste `test_formato_evento` no modelo `ordem_servico_trajeto_alternativo_sentido` para validar o formato do `evento` (sem acentos ou caracteres especiais)(https://github.com/RJ-SMTR/pipelines_v3/pull/411)
+- Adiciona o teste `test_formato_evento` no modelo `ordem_servico_trajeto_alternativo_sentido` (planejamento) e na staging `br_rj_riodejaneiro_gtfs_staging` para validar o formato do `evento` (sem acentos ou caracteres especiais)(https://github.com/RJ-SMTR/pipelines_v3/pull/411)
 
 ## [1.8.2] - 2026-07-09